Move selectors from country card to local sidebar
