+++ src/styles/settings.scss
@@ -45,6 +45,10 @@ $light-text: $white;
 $logos-text: $oslo-gray;
 $half-earth-meter-green: $screamin-green;
 $half-earth-meter-blue: $lochmara;
+$mammals-color: #92EB58;
+$amphibians-color: #9873EF;
+$birds-color: #34BD92;
+$reptiles-color: #3AA8EE;
 $priority-blue: #0664F6;
 $human-pressure-high: #282052;
 $human-pressure-mid: #9F82CE;
